refactor(register): lots of TS conversion, only track Sentry errors in prod
