+++ package.json
@@ -15,7 +15,7 @@
     "serve": "gatsby serve",
     "lint": "eslint '**/*.{js,jsx,ts,tsx}'",
     "lint:fix": "eslint '**/*.{js,jsx,ts,tsx}' --fix",
-    "type-check": "tsc --noEmit"
+    "type-check": "tsc --pretty --noEmit"
   },
   "browserslist": [
     "last 2 versions",
