+++ tsconfig.json
@@ -1,23 +1,21 @@
 {
   "compilerOptions": {
     "allowJs": true,
+    "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "lib": [
-      "dom",
-      "es5",
-      "es6"
-    ],
+    "lib": ["dom", "es2015", "es2017"],
     "module": "commonjs",
     "moduleResolution": "node",
     "noEmit": true,
     "noImplicitAny": false,
-    "noUnusedLocals": false,
+    "noUnusedLocals": true,
     "resolveJsonModule": true,
     "skipLibCheck": true,
     "sourceMap": true,
-    "target": "ES2019",
+    "strict": true,
+    "target": "es6",
   },
   "include": [
     "src/**/*"
